requirement.txt
langchain
langchain-community
langchain-google-genai
youtube-transcript-api
python-docx
beautifulsoup4
IPython
streamlit
python-dotenv
markdown
streamlit_option_menu
nltk